requirements.txt
pyautogen[websurfer,gemini]
playwright